+++ libcxx/test/libcxx/modules_include.gen.py
@@ -44,3 +44,34 @@ for header in public_headers:
 
 #include <{header}>
 """)
+
+print(f"""
+//--- __std_clang_module.compile.pass.mm
+// RUN{BLOCKLIT}: %{{cxx}} %s %{{flags}} %{{compile_flags}} -fmodules -fcxx-modules -fmodules-cache-path=%t -fsyntax-only
+
+// REQUIRES{BLOCKLIT}: modules-build
+// UNSUPPORTED{BLOCKLIT}: use_module_std
+
+// GCC doesn't support -fcxx-modules
+// UNSUPPORTED{BLOCKLIT}: gcc
+
+// The Windows headers don't appear to be compatible with modules
+// UNSUPPORTED{BLOCKLIT}: windows
+// UNSUPPORTED{BLOCKLIT}: buildhost=windows
+
+// The AIX headers don't appear to be compatible with modules
+// UNSUPPORTED{BLOCKLIT}: LIBCXX-AIX-FIXME
+
+// The Android headers don't appear to be compatible with modules yet
+// XFAIL{BLOCKLIT}: LIBCXX-ANDROID-FIXME
+
+// TODO: Investigate this failure
+// UNSUPPORTED{BLOCKLIT}: LIBCXX-FREEBSD-FIXME
+
+// Lit seems to compile this twice: once with the default flags and once with with
+// the flags specified in the RUN directive. Guard the first compile from failing.
+#if __has_feature(modules)
+@import std;
+#endif
+
+""")